added profile, login, homepage, database etc

+++ src/main/java/com/campuseats/campuseats/model/CartItem.java
@@ -1,15 +1,15 @@
 package com.campuseats.campuseats.model;
 
 import lombok.Data;
-
+import java.math.BigDecimal;
 @Data
 public class CartItem {
     private Long itemId;
     private String itemName;
-    private double price;
+    private BigDecimal price;
     private int quantity;
 
-    public CartItem(Long itemId, String itemName, double price, int quantity) {
+    public CartItem(Long itemId, String itemName, BigDecimal price, int quantity) {
         this.itemId = itemId;
         this.itemName = itemName;
         this.price = price;
